[bugfix] Transforms regular expressions to functions for compat
The `recursive-readdir` npm package expects either `glob` format or
a function to test whether a file should be included in the
recursive tree. However, the api to this package demands a
regular expression compatible pattern as of 0.8.2.

If a developer provided a regular expression exclude pattern
it would cause the package to fail (without decent logs).

Now the package will transform all patterns to regular expressions
immediately for performance, so as it does not need to generate
a new regular expression for each file path it checks, which
could be very costly if the file system (build) is large.

It will also transform the regular expression into a test function
that will be passed to `recursive-readdir`.

See: https://github.com/syndbg/webpack-google-cloud-storage-plugin/issues/16